fix: price buttons now navigates to course-detail
buying courses will be handled in the course detail page or saved courses page
after buying the course, user will be navigated to spesific course page to see the videos
