+++ app/[locale]/(content-pages)/course-detail/[courseId]/page.jsx
@@ -6,7 +6,13 @@ import { useTranslations } from "next-intl";
 import Image from "next/image";
 import { useState, useEffect } from "react";
 import Loading from "@/app/Components/LoadingPage/Loading";
-
+import { Box, Modal, Typography } from "@mui/material";
+import Backdrop from '@mui/material/Backdrop';
+import { useSession } from "next-auth/react";
+import { useLocale } from "next-intl";
+import notify from "@/utils/notifications";
+import 'react-toastify/dist/ReactToastify.css';
+import { useRouter } from "next-intl/client";
 function CourseViewImage({ imgUrl }) {
   return (
     <div className="relative w-full h-[250px] lg:h-[350px] xl:h-[400px] 2xl:h-[500px] rounded-2xl">
@@ -21,11 +27,48 @@ function CourseViewImage({ imgUrl }) {
     </div>
   );
 }
+const modalStyle = {
+  position: 'absolute',
+  top: '50%',
+  left: '50%',
+  transform: 'translate(-50%, -50%)',
+  width: 400,
+  bgcolor: '#f5f5f5',
+  border: '#2e8dff',
+  boxShadow: 24,
+  p: 4,
+  borderRadius: 2
+}
 
 export default function CourseDetail({ params: { courseId } }) {
+  const router = useRouter()
   const [selectedCourse, setSelectedCourse] = useState(null);
   const [isLoading, setIsLoading] = useState(true);
-
+  const locale = useLocale()
+  const session = useSession()
+  const [openModal, setOpenModal] = useState(false)
+  const handleOpenModal = () => setOpenModal(true)
+  const handleCloseModal = () => setOpenModal(false)
+  const handleEnroll = async (courseId) => {
+    console.log("session", session)
+    const res = await fetch("/api/course-enroll", {
+      method: "POST",
+      body: JSON.stringify({
+        courseId,
+        email: session?.data?.session?.user?.email,
+        locale
+      }),
+    })
+    if (!res.ok) {
+      const { message } = await res.json()
+      notify(message, "error")
+      return
+    }
+    const { message } = await res.json()
+    notify(message, "success")
+    handleCloseModal()
+    router.push(`/spesific-course/${courseId}`)
+  }
   useEffect(() => {
     async function fetchCourses() {
       try {
@@ -49,6 +92,49 @@ export default function CourseDetail({ params: { courseId } }) {
   const t = useTranslations("SavedCourses");
   return (
     <>
+      {/* MODAL STARTS */}
+      <Modal
+        open={openModal}
+        onClose={handleCloseModal}
+        closeAfterTransition
+        slots={{ backdrop: Backdrop }}
+        slotProps={{
+          backdrop: {
+            timeout: 500,
+          },
+        }}
+
+      >
+        <Box sx={modalStyle}>
+          <Typography id="modal-modal-title" variant="h6" component="h2">
+            {selectedCourse?.title}
+          </Typography>
+          <Typography id="modal-modal-description" sx={{ mt: 2 }}>
+            Want to enroll in this course?
+          </Typography>
+          <div className="flex justify-end gap-2 mt-4">
+            <LearnUButton
+              text="Cancel"
+              bgColor="red"
+              borderRadius={20}
+              paddingInline={25}
+              paddingBlock={5}
+              width="full"
+              onClick={handleCloseModal}
+            />
+            <LearnUButton
+
+              text="Enroll"
+              bgColor="blue"
+              borderRadius={20}
+              paddingInline={25}
+              paddingBlock={5}
+              width="full"
+              onClick={() => handleEnroll(selectedCourse?._id)}
+            />
+          </div>
+        </Box>
+      </Modal>
       {isLoading ? (
         <Loading />
       ) : (
@@ -67,6 +153,7 @@ export default function CourseDetail({ params: { courseId } }) {
             </div>
             <div className="flex max-xl:flex-col gap-4 px-8 mt-auto pb-4 lg:pb-0">
               <LearnUButton
+                onClick={handleOpenModal}
                 className={"w-full basis-full uppercase lg:mb-2"}
                 text={t("Buy Now")}
               />
